script com decimal(4,2)

+++ scriptProcessos.sql
@@ -4,14 +4,14 @@ desc processos;
 
  create table processos (
 id int primary key auto_increment,
-maceracao float,
-malteacao float,
-moagem float,
-brassagem float,
-fervura float,
-resfriamento float,
-fermentacao float,
-maturacao float,
-filtragem float,
-pasteurizacao float
+maceracao decimal(4,2),
+malteacao decimal(4,2),
+moagem decimal(4,2),
+brassagem decimal(4,2),
+fervura decimal(4,2),
+resfriamento decimal(4,2),
+fermentacao decimal(4,2),
+maturacao decimal(4,2),
+filtragem decimal(4,2),
+pasteurizacao decimal(4,2)
 );
